# apis/app-api/store/openapi.yaml
openapi: 3.1.2
info:
  title: SDKWork App Store App API
  version: 1.0.0
  description: 'App/client contract for SDKWork App Store catalog, listings, publishers,
    releases, compliance, library, and wishlist capabilities.

    '
  x-sdkwork-owner: sdkwork-appstore
  x-sdkwork-api-authority: sdkwork-appstore-app-api
  x-sdkwork-sdk-family: sdkwork-appstore-app-sdk
  x-sdkwork-domain: appstore
  x-sdkwork-standard-profile: sdkwork-appstore-app-v3
  x-sdkwork-audience: Store clients, publisher console, mobile/H5/PC apps
servers:
- url: http://127.0.0.1:18090
  description: Local sdkwork-appstore runtime
...
security:
- AuthToken: []
  AccessToken: []
paths:
...
  /app/v3/api/publishers/me/apps:
    post:
      tags:
      - publishers
      summary: Bootstrap appstore_app and draft listing for the current publisher
      operationId: appstore.publishers.me.apps.bootstrap
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublisherAppBootstrapRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/PublisherAppBootstrapResponse'
...
  /app/v3/api/listings/{listingId}/compliance:
    get:
      tags:
      - compliance
      summary: Retrieve compliance profile
      operationId: appstore.compliance.profile.retrieve
      parameters:
      - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/ComplianceProfileResponse'
    put:
      tags:
      - compliance
      summary: Update compliance profile
      operationId: appstore.compliance.profile.update
      parameters:
      - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceProfileUpdateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/ComplianceProfileResponse'
  /app/v3/api/listings/{listingId}/compliance/permissions:
    put:
      tags:
      - compliance
      summary: Update permission disclosures
      operationId: appstore.compliance.permissions.update
      parameters:
      - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompliancePermissionUpdateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/CompliancePermissionListResponse'
  /app/v3/api/listings/{listingId}/compliance/iap_items:
    get:
      tags:
      - compliance
      summary: List in-app purchase items for compliance
      operationId: appstore.compliance.iapItems.list
      parameters:
      - $ref: '#/components/parameters/ListingId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdkWorkListResponse'
...
  /app/v3/api/library/install:
    post:
      tags:
      - library
      summary: Record install and prepare download
      operationId: appstore.library.install
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryInstallRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SdkWorkApiResponse'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
                      required:
                      - item
                      properties:
                        item:
                          $ref: '#/components/schemas/LibraryInstallResponse'
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: Auth-Token
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    ListingId:
      name: listingId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    MediaResource:
      type: object
      required:
      - id
      - kind
      properties:
        id:
          type: string
        kind:
          type: string
        url:
          type: string
        driveNodeId:
          type: string
    ListingSummary:
      type: object
      required:
      - id
      - appKey
      - displayName
      - listingSlug
      - pricingModel
      properties:
        id:
          type: string
        appId:
          type: string
        appKey:
          type: string
        displayName:
          type: string
        subtitle:
          type: string
        listingSlug:
          type: string
        pricingModel:
          type: string
          enum:
          - FREE
          - PAID
          - FREEMIUM
          - SUBSCRIPTION
        icon:
          $ref: '#/components/schemas/MediaResource'
        averageRating:
          type: string
        ratingCount:
          type: integer
...
    PublisherAppBootstrapRequest:
      type: object
      required:
      - appKey
      - displayName
      - defaultLocale
      properties:
        appKey:
          type: string
        displayName:
          type: string
        defaultLocale:
          type: string
        appType:
          type: string
        listingSlug:
          type: string
        pricingModel:
          type: string
    PublisherAppBootstrapResponse:
      type: object
      required:
      - app
      - listing
      properties:
        app:
          $ref: '#/components/schemas/StoreAppSummary'
        listing:
          $ref: '#/components/schemas/ListingSummary'
    StoreAppSummary:
      type: object
      required:
      - id
      - appKey
      - displayName
      properties:
        id:
          type: string
        appKey:
          type: string
        appSlug:
          type: string
        displayName:
          type: string
        defaultLocale:
          type: string
        appType:
          type: string
        appStatus:
          type: string
        publisherId:
          type: string
        currentListingId:
          type: string
...
    ComplianceProfileResponse:
      $ref: '#/components/schemas/SdkWorkResourceResponse'
    ComplianceProfileUpdateRequest:
      type: object
      properties:
        privacyNutrition:
          type: object
          additionalProperties: true
        contentRatingQuestionnaire:
          type: object
          additionalProperties: true
        dataSafety:
          type: object
          additionalProperties: true
    CompliancePermissionUpdateRequest:
      type: object
      required:
      - permissions
      properties:
        permissions:
          type: array
          items:
            type: object
            required:
            - permissionCode
            - usagePurpose
            properties:
              permissionCode:
                type: string
              usagePurpose:
                type: string
              isRequired:
                type: boolean
    CompliancePermissionListResponse:
      $ref: '#/components/schemas/SdkWorkListResponse'
...
    LibraryInstallRequest:
      type: object
      required:
      - listingId
      - platform
      properties:
        listingId:
          type: string
        platform:
          type: string
        architecture:
          type: string
        deviceId:
          type: string
    LibraryInstallResponse:
      $ref: '#/components/schemas/SdkWorkResourceResponse'
...
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - data
      - traceId
      properties:
        code:
          type: integer
          format: int32
          enum:
          - 0
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload typed per response schema.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
    SdkWorkResourceData:
      type: object
      additionalProperties: false
      required:
      - item
      properties:
        item:
          type: object
          additionalProperties: true
          description: Typed domain resource for the operation.
    SdkWorkPageData:
      type: object
      additionalProperties: false
      required:
      - items
      - pageInfo
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    PageInfo:
      type: object
      additionalProperties: false
      required:
      - mode
      properties:
        mode:
          type: string
          enum:
          - offset
          - cursor
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: string
          pattern: ^[0-9]+$
        totalPages:
          type: integer
          minimum: 0
        nextCursor:
          type:
          - string
          - 'null'
        hasMore:
          type: boolean
...
    SdkWorkResourceResponse:
      allOf:
      - $ref: '#/components/schemas/SdkWorkApiResponse'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/SdkWorkResourceData'
    SdkWorkListResponse:
      allOf:
      - $ref: '#/components/schemas/SdkWorkApiResponse'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/SdkWorkPageData'
...